requirements.txt
Django
django-crispy-forms
pytube3
youtube-transcript-api
moviepy
pytube
requests
openai-whisper
transformers
sentencepiece
TTS
inflect
ffmpeg-python
pydub